backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.136.1",
    "python-dotenv>=1.2.2",
    "python-multipart>=0.0.26",
    "stripe>=15.1.0",
    "uvicorn>=0.46.0",
]
